test: added tests for message route

+++ app/Controllers/Http/Authenticated/MessagesController.ts
@@ -3,6 +3,7 @@ import { rules, schema } from '@ioc:Adonis/Core/Validator'
 import Database from '@ioc:Adonis/Lucid/Database'
 import { Conversation, Message, User } from 'App/Models'
 import { StoreValidator } from 'App/Validators/Authenticated/Message'
+import { DateTime } from 'luxon'
 
 export default class MessagesController {
   public async store({ auth, request, response }: HttpContextContract) {
@@ -26,7 +27,18 @@ export default class MessagesController {
     const guest = await User.findByOrFail('username', username)
     const conversation = await Conversation.create({ ownerId: auth.user!.id, guestId: guest.id })
 
-    return await conversation.related('messages').create({ userId: auth.user!.id, content })
+    const message = await conversation
+      .related('messages')
+      .create({ userId: auth.user!.id, content })
+
+    await message.load('user')
+
+    return message.serialize({
+      fields: { omit: ['conversationId'] },
+      relations: {
+        user: { fields: { pick: ['username', 'email'] } }
+      }
+    })
   }
 
   public async show({ request }: HttpContextContract) {
@@ -46,42 +58,46 @@ export default class MessagesController {
     return messages
   }
 
-  public async update({ auth, request }: HttpContextContract) {
-    const { id, content } = await request.validate({
-      schema: schema.create({
-        id: schema.number([
-          rules.trim(),
-          rules.exists({ column: 'id', table: 'messages', where: { user_id: auth.user!.id } })
-        ]),
-        content: schema.string([rules.trim()])
-      })
-    })
+  // public async update({ auth, request }: HttpContextContract) {
+  //   const { id, content } = await request.validate({
+  //     schema: schema.create({
+  //       id: schema.number([
+  //         rules.trim(),
+  //         rules.exists({ column: 'id', table: 'messages', where: { user_id: auth.user!.id } })
+  //       ]),
+  //       content: schema.string([rules.trim()])
+  //     })
+  //   })
 
-    const message = await Message.findOrFail(id)
+  //   const message = await Message.findOrFail(id)
 
-    await message.merge({ content }).save()
+  //   await message.merge({ content }).save()
 
-    await message.load('user')
+  //   await message.load('user')
 
-    return message.serialize({
-      fields: { omit: ['conversationId'] },
-      relations: {
-        user: { fields: { pick: ['username', 'email'] } }
-      }
-    })
-  }
+  //   return message.serialize({
+  //     fields: { omit: ['conversationId'] },
+  //     relations: {
+  //       user: { fields: { pick: ['username', 'email'] } }
+  //     }
+  //   })
+  // }
 
-  public async destroy({ auth, request }: HttpContextContract) {
+  public async destroy({ auth, request, response }: HttpContextContract) {
     const { id } = await request.validate({
       schema: schema.create({
         id: schema.number([
           rules.trim(),
-          rules.exists({ column: 'id', table: 'messages', where: { userId: auth.user!.id } })
+          rules.exists({ column: 'id', table: 'messages', where: { user_id: auth.user!.id } })
         ])
       })
     })
 
     const message = await Message.findOrFail(id)
+
+    if (message.createdAt > DateTime.now().plus({ minute: 30 })) {
+      return response.badRequest()
+    }
 
     await message.delete()
   }
